Obsolete python-matplotlib-wx for clean updates.
The python-matplotlib-wx package was only two python files (sans pyc & pyo),
that actually now are merged in the main package. This is a packaging issue
that should be corrected by also adding extra backend subpackages, but those
need extra packages in fedora. There are actually several other backends
that were in the main package, but the ones with .so files are split (main
package and tk backend).